k3.cc

+++ gentle/language_model.py
@@ -9,11 +9,13 @@ import tempfile
 from .util.paths import get_binary
 from .metasentence import MetaSentence
 from .resources import Resources
+from kaldi_model import kaldi_lm
 
 MKGRAPH_PATH = get_binary("ext/m3")
 
 # [oov] no longer in words.txt
 OOV_TERM = '<unk>'
+
 
 def make_bigram_lm_fst(word_sequences, **kwargs):
     '''
@@ -43,7 +45,7 @@ def make_bigram_lm_fst(word_sequences, **kwargs):
             continue
 
         prev_word = word_sequence[0]
-        bigrams[OOV_TERM].add(prev_word) # valid start (?)
+        bigrams[OOV_TERM].add(prev_word)  # valid start (?)
 
         if disfluency:
             bigrams[OOV_TERM].update(disfluencies)
@@ -70,6 +72,7 @@ def make_bigram_lm_fst(word_sequences, **kwargs):
         bigrams.setdefault(prev_word, set()).add(OOV_TERM)
 
     node_ids = {}
+
     def get_node_id(word):
         node_id = node_ids.get(word, len(node_ids) + 1)
         node_ids[word] = node_id
@@ -94,6 +97,7 @@ def make_bigram_lm_fst(word_sequences, **kwargs):
 
     return output.encode()
 
+
 def make_bigram_language_model(kaldi_seq, proto_langdir, **kwargs):
     """Generates a language model to fit the text.
 
@@ -111,24 +115,24 @@ def make_bigram_language_model(kaldi_seq, proto_langdir, **kwargs):
     txt_fst_file.close()
 
     hclg_filename = tempfile.mktemp(suffix='_HCLG.fst')
+
+    lm = kaldi_lm()
+    status = 0
     try:
-        devnull = open(os.devnull, 'wb')
-        subprocess.check_output([MKGRAPH_PATH,
-                        proto_langdir,
-                        txt_fst_file.name,
-                        hclg_filename],
-                        stderr=devnull)
+        status = lm.make_lm(proto_langdir, txt_fst_file.name, hclg_filename)
     except Exception as e:
-        try:
-            os.unlink(hclg_filename)
-        except:
-            pass
+        if os.path.exists(txt_fst_file.name):
+            os.unlink(txt_fst_file.name)
         raise e
-    finally:
+    if os.path.exists(txt_fst_file.name):
         os.unlink(txt_fst_file.name)
+
+    if status != 1:
+        raise Exception("Unable to create fst file. Check logs for more details.")
 
     return hclg_filename
 
-if __name__=='__main__':
+
+if __name__ == '__main__':
     import sys
     make_bigram_language_model(open(sys.argv[1]).read(), Resources().proto_langdir)
